+++ packages/filecoin-api/src/aggregator/service.js
@@ -77,6 +77,7 @@ export const pieceAccept = async ({ capability }, context) => {
     }
   }
 
+  // Get buffered pieces 
   const [{ aggregate, inclusion }] = getInclusionRes.ok
   const getAggregateRes = await context.aggregateStore.get({ aggregate })
   if (getAggregateRes.error) {
@@ -96,6 +97,7 @@ export const pieceAccept = async ({ capability }, context) => {
         aggregate,
         pieces,
       },
+      expiration: Infinity,
     })
     .delegate()